+++ vcsn/misc/fwd.hh
@@ -7,4 +7,10 @@ namespace vcsn
 
   // misc/wet.hh
   enum class wet_kind_t;
+
+  namespace rat
+  {
+    class location;
+    class position;
+  };
 }